If statement added inside loop removed an error

+++ Latex_Script.sh
@@ -20,7 +20,12 @@ i=1 # Sets i to 1
 	     until [ "$text" = "" ] # When this condition is met the loop has reached the end of the Unwanted_Strings.txt file and the loop exits
            do	 
              eval 'sed '$text''
-             mv tempfile.txt? inputCopy.txt               
+			   if [ -e "tempfile.txt" ]
+			     then
+                 mv tempfile.txt inputCopy.txt 
+                 else
+				 mv tempfile.txt? inputCopy.txt 
+               fi             
 	         i=$(($i+1)) # increments i by 1
 		     text=$(sed -n ''$i'{p;q}' Unwanted_Strings.txt)  # Sets text to the next string in the list in Unwanted_Strings.txt
          done
